ODataScrapperBPM/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pyodbc==4.0.27
PyYAML==5.2
requests==2.22.0
urllib3==1.25.7
